requirements.txt
requests>=2.31.0
pandas>=2.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
pymongo>=4.5.0
tqdm>=4.65.0
pydantic>=2.0.0
httpx>=0.24.0
aiohttp>=3.8.0
asyncio-throttle>=1.0.2

# PostgreSQL
psycopg2-binary>=2.9.9

# Airflow (for local development - use docker for production)
apache-airflow>=2.8.0
pendulum>=3.0.0

